Bugfix in Chapter 6.1
Fixed some copy and paste errors in sandbox code at the bottom of the
page which prevented the example to run.

I did test the code on my local machine after fixing it, but the sandbox
did not run on my local machine using mdbook serve.

And I apologize for various new lines, these are added by the prettier
plugin of my editor.